adding project_skill relation

+++ db/schema.rb
@@ -15,8 +15,8 @@ ActiveRecord::Schema[7.0].define(version: 2022_07_06_150447) do
   enable_extension "plpgsql"
 
   create_table "project_skills", force: :cascade do |t|
-    t.integer "project_id", null: false
-    t.integer "skill_id", null: false
+    t.bigint "project_id", null: false
+    t.bigint "skill_id", null: false
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
     t.index ["project_id"], name: "index_project_skills_on_project_id"
